prepare data import with files job
