+++ server/modules/analysis/service/rChain/PersistResultsJob.js
@@ -3,7 +3,7 @@ import * as NodeDef from '@core/survey/nodeDef'
 import * as Record from '@core/record/record'
 
 import DataImportJob from '@server/modules/dataImport/service/DataImportJob'
-import { DataImportFileReader } from '@server/modules/dataImport/service/DataImportJob/dataImportFileReader'
+import { DataImportCsvFileReader } from '@server/modules/dataImport/service/DataImportJob/dataImportCsvFileReader'
 import * as SurveyManager from '@server/modules/survey/manager/surveyManager'
 import FileZip from '@server/utils/file/fileZip'
 
@@ -47,7 +47,7 @@ export default class PersistResultsJob extends DataImportJob {
   async createCsvReader() {
     const { cycle, nodeDefUuid, survey } = this.context
 
-    return DataImportFileReader.createReaderFromStream({
+    return DataImportCsvFileReader.createReaderFromStream({
       stream: this.stream,
       survey,
       cycle,
