+++ server/modules/dataImport/service/DataImportJob/DataFilesImportJob.js
@@ -1,0 +1,11 @@
+import FilesImportJob from '@server/modules/arenaImport/service/arenaImport/jobs/filesImportJob'
+
+export default class DataFilesImportJob extends FilesImportJob {
+  constructor(params) {
+    super(params)
+  }
+
+  async fetchFilesSummaries() {}
+
+  async fetchFileContent({ fileName, fileUuid }) {}
+}
